arbitration and blocking functions moved from enhanced to the bus
