+++ include/bus.hpp
@@ -1,9 +1,31 @@
 #ifndef _BUS_H_
 #define _BUS_H_
-#include "main.hpp"
+#include <queue>
+#include <WiFiClient.h>
 #include "busstate.hpp"
 #include "arbitration.hpp"
-#include "queue"
+
+enum responses {
+    RESETTED = 0x0,
+    RECEIVED = 0x1,
+    STARTED = 0x2,
+    INFO = 0x3,
+    FAILED = 0xa,
+    ERROR_EBUS = 0xb,
+    ERROR_HOST = 0xc
+};
+
+enum errors {
+    ERR_FRAMING = 0x00,
+    ERR_OVERRUN = 0x01
+};
+
+void getArbitrationClient(WiFiClient* &client, uint8_t &address);
+void clearArbitrationClient();
+bool setArbitrationClient(WiFiClient* &client, uint8_t &address);
+
+void arbitrationDone();
+WiFiClient* arbitrationRequested(uint8_t& arbitration_client);
 
 #ifdef ESP32
 #include "atomic"
